requirements.txt
streamlit
pandas
matplotlib
plotly
gspread
google-auth
python-dotenv
langchain
langchain-google-genai
psycopg2
psycopg2-binary
passlib
bcrypt==4.3.0